+++ src/libvfplot/gstate.h
@@ -9,6 +9,8 @@
 #ifndef GSTATE_H
 #define GSTATE_H
 
+#include <stdlib.h>
+
 /*
   this reads and writes the collection of arrows and
   network neighbours to/from a file. The read function
@@ -22,19 +24,19 @@
 typedef struct {
   struct
   {
-    int n;
+    size_t n;
     arrow_t *A;
   } arrow;
   struct
   {
-    int n;
+    size_t n;
     nbs_t* N;
   } nbs;
 } gstate_t;
 
-#define GSTATE_NULL {{0,NULL},{0,NULL}}
+#define GSTATE_NULL {{0, NULL}, {0, NULL}}
 
-extern int gstate_read(char*,gstate_t*);
-extern int gstate_write(char*,gstate_t*);
+extern int gstate_read(char*, gstate_t*);
+extern int gstate_write(char*, gstate_t*);
 
 #endif
