+++ src/libvfplot/hedgehog.h
@@ -10,6 +10,7 @@
 
 #include <vfplot/vfplot.h>
 
-extern int vfplot_hedgehog(domain_t*,vfun_t,cfun_t,void*,vfp_opt_t,int*,arrow_t**);
+extern int vfplot_hedgehog(domain_t*, vfun_t, cfun_t,
+			   void*, vfp_opt_t, size_t*, arrow_t**);
 
 #endif
